requirements.txt
accelerate==0.28.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
backoff==2.2.1
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
deepdiff==6.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
einops==0.7.0
emoji==2.11.0
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
filetype==1.2.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gradio==4.24.0
gradio_client==0.14.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.20.3
idna==3.6
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.3
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langdetect==1.0.9
llama-index==0.10.26
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.26
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-mistralai==0.1.3
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.1.4
llama-index-llms-mistralai==0.1.4
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-milvus==0.1.6
llama-index-vector-stores-pinecone==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.15
lxml==5.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.5
mdurl==0.1.2
milvus==2.3.5
minijinja==1.0.16
minio==7.2.5
mistralai==0.0.11
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openai==1.14.3
openapi==1.1.0
ordered-set==4.1.0
orjson==3.10.0
packaging==24.0
pandas==2.0.3
pathspec==0.12.1
pillow==10.2.0
pinecone-client==3.2.2
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
protobuf==5.26.1
psutil==5.9.8
pyarrow==15.0.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pymilvus==2.4.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
pypdf==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.7.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.4
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.6.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
